moved historycards to its own component, images to public images dir, removed old toast components
