+++ client/src/components/Loading/HistoryLoading.jsx
@@ -1,12 +1,16 @@
-const PageLoading = (props) => {
+const PageLoading = () => {
     return (
         <div className="loading-ctn container">
             <h1>Loading Content</h1>
             <div className="flex flex-col gap-4 w-52">
-                <div className="skeleton h-32 w-full"></div>
-                <div className="skeleton h-4 w-28"></div>
-                <div className="skeleton h-4 w-full"></div>
-                <div className="skeleton h-4 w-full"></div>
+                <div className="skeleton w-[400px] h-[250px] rounded-2xl"></div>
+                <div className="skeleton w-[400px] h-[250px] rounded-2xl"></div>
+                <div className="skeleton w-[400px] h-[250px] rounded-2xl"></div>
+                <div className="skeleton w-[400px] h-[250px] rounded-2xl"></div>
+                <div className="skeleton w-[400px] h-[250px] rounded-2xl"></div>
+                <div className="skeleton w-[400px] h-[250px] rounded-2xl"></div>
+                <div className="skeleton w-[400px] h-[250px] rounded-2xl"></div>
+                <div className="skeleton w-[400px] h-[250px] rounded-2xl"></div>
             </div>
         </div>
